+++ zest/releaser/postrelease.py
@@ -151,6 +151,7 @@ class Postreleaser(baserelease.Basereleaser):
             logger.info("The '%s':\n\n%s\n" % (diff_cmd, diff))
         if utils.ask("OK to commit this"):
             msg = self.data['commit_msg'] % self.data
+            msg = self.update_commit_message(msg)
             commit_cmd = self.vcs.cmd_commit(msg)
             commit = system(commit_cmd)
             logger.info(commit)