+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog for zest.releaser
 -----------------
 
 - Add optional support for uploading Python wheels.  Use the new
-  ``zest.releaser[recommended]`` extra, or pip install ``wheel``
+  ``zest.releaser[recommended]`` extra, or run ``pip install wheel``
   yourself next to ``zest.releaser``.  Create or edit ``setup.cfg`` in
   your project (or globally in your ``~/.pypirc``) and create a section
   ``[zest.releaser]`` with ``create_wheel = yes`` to create a wheel to