+++ CHANGES.rst
@@ -8,7 +8,7 @@ Changelog for zest.releaser
   ``zest.releaser[recommended]`` extra, or run ``pip install wheel``
   yourself next to ``zest.releaser``.  Create or edit ``setup.cfg`` in
   your project (or globally in your ``~/.pypirc``) and create a section
-  ``[zest.releaser]`` with ``create_wheel = yes`` to create a wheel to
+  ``[zest.releaser]`` with ``create-wheel = yes`` to create a wheel to
   upload to PyPI.  See http://pythonwheels.com for deciding whether
   this is a good idea for your package.  Briefly, if it is a pure
   Python 2 *or* pure Python 3 package: just do it.